TeamHR/requirement.txt
astroid==2.4.1
beautifulsoup4==4.9.1
boto3==1.14.4
botocore==1.17.4
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==2.2
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
flake8==3.8.2
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
ngrok==0.0.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pydocstyle==4.0.1
pyflakes==2.2.0
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
